User login, logout and sign up finished

+++ studybud/base/views.py
@@ -1,11 +1,32 @@
 from django.shortcuts import render, redirect
+from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
+from django.http import HttpResponse
 
 from .forms import RoomForm
-from .models import Room
+from .models import Room, Topic
 
 
 def home(request):
-    context = {'rooms': Room.objects.all()}
+    q = request.GET.get('q') if request.GET.get('q') else ''
+
+    rooms = Room.objects.filter(
+        Q(topic__name__icontains=q) |
+        Q(name__icontains=q) |
+        Q(description__icontains=q) |
+        Q(host__username__icontains=q)
+    )
+
+    context = {
+        'rooms': rooms,
+        'room_count': rooms.count(),
+        'topics': Topic.objects.all()
+    }
+
     return render(request, 'base/home.html', context)
 
 
@@ -14,11 +35,13 @@ def room(request, pk):
     return render(request, 'base/room.html', context)
 
 
+@login_required(login_url='login')
 def create_room(request):
     form = RoomForm()
 
     if request.method == 'POST':
         form = RoomForm(request.POST)
+
         if form.is_valid():
             form.save()
             return redirect('home')
@@ -27,12 +50,17 @@ def create_room(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
 
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
+        
         if form.is_valid():
             form.save()
             return redirect('home')
@@ -41,11 +69,70 @@ def update_room(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
     
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     
     return render(request, 'base/delete.html', {'obj': room})
+
+
+def login_page(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        print(username, password)
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')
+        
+        user = authenticate(request, username=username, password=password)
+        
+        if user:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or password is incorrect')
+
+    context = {'page': 'login'}
+    return render(request, 'base/login_register.html', context)
+
+
+def logout_page(request):
+    logout(request)
+    return redirect('home')
+
+
+def register_page(request):
+    form = UserCreationForm()
+    
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            
+            login(request, user)
+
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occurred during registration')
+            
+    context = {'form': form}
+    return render(request, 'base/login_register.html', context)
+
+
